+++ shinnagoing/Controller/LoginVC.swift
@@ -123,30 +123,22 @@ class LoginVC: UIViewController {
         let vc = JoinVC()
         self.navigationItem.backButtonTitle = ""
         navigationController?.pushViewController(vc, animated: true)
-}
-    
     }
-    @objc func buttonTapped() {
-        print("Tap")
-        let vc = JoinVC()
-        self.navigationItem.backButtonTitle = ""
-        navigationController?.pushViewController(vc, animated: true)
     
     
 }
-
 extension UIColor {
     convenience init(hex: String, alpha: CGFloat = 1.0) {
         var hexSanitized = hex.trimmingCharacters(in: .whitespacesAndNewlines)
         hexSanitized = hexSanitized.replacingOccurrences(of: "#", with: "")
-
+        
         var rgb: UInt64 = 0
         Scanner(string: hexSanitized).scanHexInt64(&rgb)
-
+        
         let red = CGFloat((rgb & 0xFF0000) >> 16) / 255.0
         let green = CGFloat((rgb & 0x00FF00) >> 8) / 255.0
         let blue = CGFloat(rgb & 0x0000FF) / 255.0
-
+        
         self.init(red: red, green: green, blue: blue, alpha: 1.0)
     }
 }
