+++ shinnagoing/Controller/MapScreen/ModalVC.swift
@@ -29,9 +29,12 @@ class ModalVC: UIViewController {
         batteryLabel.textColor = .black
         batteryLabel.textAlignment = .center
         
-        let batteryImageView = UIImageView(image: UIImage(named: "battery"))
+        let batteryImageView = UIImageView()
+        batteryImageView.image = imageForBatteryLevel(Int(battery))
         batteryImageView.contentMode = .scaleAspectFit
-        
+        batteryImageView.tintColor = UIColor(hex: "915B5B")
+
+
         let kickBoardImageView = UIImageView(image: UIImage(named: "kickBoard"))
         kickBoardImageView.contentMode = .scaleAspectFit
         
@@ -75,6 +78,19 @@ class ModalVC: UIViewController {
             rentButton
         ].forEach { view.addSubview($0) }
         
+        func imageForBatteryLevel(_ battery: Int) -> UIImage? {
+            switch battery {
+            case 70...100:
+                return UIImage(systemName: "battery.100")  // 🔋 3칸 이미지
+            case 30...69:
+                return UIImage(systemName: "battery.50")  // 🔋 2칸 이미지
+            case 0...29:
+                return UIImage(systemName: "battery.25")  // 🔋 1칸 이미지
+            default:
+                return UIImage(named: "battery.0")  // 예외처리 이미지
+            }
+        }
+        
         // --- 오토레이아웃 ---
         batteryLabel.snp.makeConstraints {
             $0.top.equalTo(view.safeAreaLayoutGuide).offset(20)
@@ -83,9 +99,9 @@ class ModalVC: UIViewController {
         
         batteryImageView.snp.makeConstraints {
             $0.centerY.equalTo(batteryLabel)
-            $0.leading.equalTo(batteryLabel.snp.trailing).offset(8)
-            $0.height.equalTo(29)
-            $0.width.equalTo(55)
+            $0.leading.equalTo(batteryLabel.snp.trailing).offset(3)
+            $0.width.equalTo(50)
+            $0.height.equalTo(30)
         }
         
         kickBoardImageView.snp.makeConstraints { make in
@@ -94,7 +110,7 @@ class ModalVC: UIViewController {
             make.leading.equalTo(batteryImageView.snp.trailing).offset(145)
             make.top.equalToSuperview().offset(17)
         }
-        
+    
         priceLabel.snp.makeConstraints {
             $0.top.equalTo(batteryLabel.snp.bottom).offset(1)
             $0.leading.equalTo(view.safeAreaLayoutGuide).offset(30)
